backend/requirements.txt
Django>=4.2,<6.1
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
redis
groq
python-dotenv
xgboost>=2.0.3
shap>=0.44.1
scikit-learn>=1.4.0
pandas>=2.2.0
